backend/requirements.txt
flask
PyMuPDF           
beautifulsoup4      
sentence-transformers
faiss-cpu
requests
python-dotenv
transformers
PyPDF2